server/requirements.txt
websockets==10.4
fastmcp>=0.4.1
python-dotenv==1.0.0
